chore: extract Settings.IndexSinkSettings into toplevel
